+++ app/Views/users/list.php
@@ -1,0 +1,9 @@
+<?= $this->extend('layouts/template') ?>
+
+<?= $this->section('content') ?>
+<div class="row">
+    <div class="col-lg-12">
+    
+    </div>
+</div>
+<?= $this->endSection() ?>